refactor: corner 설정 util 파일로 생성
