+++ pennyway-client-iOS/pennyway-client-iOS/Utils/View/RoundedCornerUtil.swift
@@ -1,0 +1,14 @@
+
+import SwiftUI
+
+// MARK: - RoundedCornerUtil
+
+struct RoundedCornerUtil: Shape {
+    var radius: CGFloat = .infinity
+    var corners: UIRectCorner = .allCorners
+
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
+        return Path(path.cgPath)
+    }
+}
